hpcgap/src/objects.c: sync
